1509. Minimum Difference Between Largest and Smallest Value in Three Moves.
